# backend/src/swagger.yaml
openapi: 3.0.3
info:
  title: CapstoneHub API
  version: 1.0.0
paths:
...
  /api/messages:
    get:
      summary: List threaded messages for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Message list
    post:
      summary: Send a message to another user
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created message
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT